move premium into client component
